requirements.txt
flask
python-dotenv
google-genai
azure-cognitiveservices-speech
gunicorn